Move import to flat routes
