+++ packages/insomnia/src/ui/routes/import.resources.tsx
@@ -1,0 +1,82 @@
+import type { ActionFunctionArgs } from 'react-router';
+
+import { importResourcesToProject, importResourcesToWorkspace } from '../../common/import';
+import * as models from '../../models';
+import { isRemoteProject } from '../../models/project';
+import type { Workspace } from '../../models/workspace';
+import {
+  initializeLocalBackendProjectAndMarkForSync,
+  pushSnapshotOnInitialize,
+} from '../../sync/vcs/initialize-backend-project';
+import { VCSInstance } from '../../sync/vcs/insomnia-sync';
+import { invariant } from '../../utils/invariant';
+import { fetchAndCacheOrganizationStorageRule } from './organization';
+
+export interface ImportResourcesActionResult {
+  errors?: string[];
+  done: boolean;
+}
+
+export async function action({ request }: ActionFunctionArgs) {
+  const formData = await request.formData();
+  const organizationId = formData.get('organizationId');
+  const projectId = formData.get('projectId');
+  const workspaceId = formData.get('workspaceId');
+
+  invariant(typeof organizationId === 'string', 'OrganizationId is required.');
+  invariant(typeof projectId === 'string', 'ProjectId is required.');
+
+  const project = await models.project.getById(projectId);
+  invariant(project, 'Project not found.');
+  if (typeof workspaceId === 'string' && workspaceId) {
+    await importResourcesToWorkspace({
+      workspaceId: workspaceId,
+    });
+
+    // TODO: find more elegant way to wait for import to finish
+    return { done: true };
+  }
+
+  await importResourcesToProject({
+    projectId: project._id,
+    syncNewWorkspaceIfNeeded,
+  });
+  return { done: true };
+}
+
+// The reason why we put this function here is because this function indirectly depends on some modules that can only run in a browser environment.
+// If we put this function in import.ts which is depended by Inso CLI, Inso CLI will fail to build because it doesn't have access to the browser environment.
+// So we put this function here and pass it to importResourcesToProject func to avoid the dependency issue.
+async function syncNewWorkspaceIfNeeded(newWorkspace: Workspace) {
+  const project = await models.project.getById(newWorkspace.parentId);
+  invariant(project, 'Project not found');
+  const userSession = await models.userSession.getOrCreate();
+
+  if (userSession.id && isRemoteProject(project)) {
+    const storageRules = await fetchAndCacheOrganizationStorageRule(project.parentId);
+    invariant(storageRules, 'Storage rules not found');
+
+    if (storageRules.enableCloudSync) {
+      // Create default env, cookie jar, and meta
+      await models.environment.getOrCreateForParentId(newWorkspace._id);
+      await models.cookieJar.getOrCreateForParentId(newWorkspace._id);
+      await models.workspaceMeta.getOrCreateByParentId(newWorkspace._id);
+      try {
+        const vcs = VCSInstance().newInstance();
+        await initializeLocalBackendProjectAndMarkForSync({
+          vcs,
+          workspace: newWorkspace,
+        });
+        await pushSnapshotOnInitialize({
+          vcs,
+          workspace: newWorkspace,
+          project,
+        });
+      } catch (e) {
+        console.warn(
+          `Failed to initialize sync to insomnia cloud for workspace ${newWorkspace._id}. This will be retried when the workspace is opened on the app. ${e.message}`,
+        );
+      }
+    }
+  }
+}
